+++ src/app/create/[gameId]/attributes.tsx
