+++ src/app/create/styles.css
@@ -1,0 +1,42 @@
+.container {
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    height: 100vh;
+    width: 100vw;
+    max-width: 100%;
+}
+
+.carousel {
+    margin-bottom: 100px;
+}
+
+.card {
+    background-size: cover;
+    background-position: center;
+}
+
+.title {
+    font-size: 35px;
+    margin-bottom: 20px;
+}
+
+.subtitle {
+    font-size: 30px;
+    margin-bottom: 20px;
+    margin-top: 100px;
+}
+
+.button {
+    background-color: grey;
+    color: white;
+    border-radius: 20px;
+    font-size: 20;
+}
+
+.text {
+    background-color: white;
+    border-radius: 25px;
+    padding: 10px;
+}
